Rainbow
Created a script that changes the background color to a color in an
array when a button is clicked.  The button is a variable defined by
getElementbyId.  The rainbow is a variable created from an array of 7
colours.

A function is created that changes the background style by calling the
variable rainbow and selecting a random number using Math object to
select a color from the array.

An event handler is created to keep track of when the button is clicked.

I sort of understand how it is working but will continue learning and
further understanding how it all ties together.